pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-lab"
version = "0.1.0"
description = "Reusable components for LLM pre-training and post-training experiments."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "LLM Lab Team", email = "ml@example.com"},
]
keywords = ["llm", "machine-learning", "pytorch", "transformers", "hydra"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Unix",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "accelerate>=0.20",
    "datasets>=2.15",
    "hydra-core>=1.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "pydantic>=2.6",
    "torch>=2.1",
    "transformers>=4.35",
    "typer>=0.12",
    "wandb>=0.16",
]

[project.optional-dependencies]
dev = [
    "black>=23.11",
    "isort>=5.12",
    "mypy>=1.8",
    "pre-commit>=3.5",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "types-PyYAML>=6.0",
]
notebooks = [
    "ipykernel>=6.25",
    "jupyter>=1.0",
]

[project.urls]
Homepage = "https://github.com/pleiadian53/llm-lab"
Documentation = "https://github.com/pleiadian53/llm-lab"
Issues = "https://github.com/pleiadian53/llm-lab/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["llm_lab"]
known_third_party = ["datasets", "hydra_core", "numpy", "pandas", "pydantic", "torch", "transformers", "typer", "wandb"]
combine_as_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",
    "F",
    "I",
    "UP",
    "N",
    "B",
    "C4",
    "TID",
]
ignore = ["E203", "E501"]

[tool.ruff.isort]
known-first-party = ["llm_lab"]
combine-as-imports = true

[tool.pytest.ini_options]
addopts = "--strict-markers --disable-warnings --maxfail=1"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
ignore_missing_imports = true
